Implement OwnedObject.create()
The implementation itself is easy. This requires a breaking change to
the OwnedObject API though. Fortunately, OwnedObject is not a public
API.

Added an argument in position three for a builder function that will
accept a single string argument and construct an instanace appropriate
to this OwnedObject container. The builder function is used to
implement the OwnedObject.create() method.