Fix a circular import issue
The wildcard import isn't working in validation.py because
identified.py imports validation. Use duck typing to work around the
problem. Look for an `identity` attribute instead of testing for
isinstance(Identified). Since the identity property is what needs to
be tested we don't have to care whether the object is an Identified,
just that it has an 'identity' attribute.