Make type_uri an optional keyword argument
For Collection and Implementation, make type_uri an optional keyword
argument instead of having type_uri be the first constructor
argument. Typical backward compatible usage is to pass URI
first. type_uri governs storage in the parent object, so make it clear
to the user that they are changing it.