Return str instead of URIRef
Change URIProperty and ReferencedObject to return
str instead of URIRef. This makes the module
compatible with that aspect of pySBOL.

This is a big change, requiring updates to
internals that compare `identity` to a URIRef,
which percolates deep in the document code for
reading RDF and for finding objects.

This also required many updates to unit tests
that had been checking for URIRefs returned from
URIProperty and from ReferencedObject.