Start fixing SBOLError argument order
The original pySBOL2 code changed the order of arguments to SBOLError.
When code was ported from pySBOL it has arguments in a different order.
A patch fix was put in place to detect arguments out of order, issue
a runtime warning, and fix the order. That left a mix of argument
ordering.

Now we are migrating to code that has the arguments in the same order
as pySBOL. This first wave fixes all the SBOLError calls that are part
of the unit tests.